requirements.txt
streamlit
pandas
python-jobspy
duckduckgo-search
sentence-transformers
google-generativeai
pypdf
python-dotenv
openpyxl
